fixed problem when moving string created using single quote, instead of double quote to arb file. It required replacing first and last char in arbValue with correct quotes
